feat: extract ranking data models for fusion and embedding space tools
