+++ services/agent/src/vss_agents/tools/fusion.py
@@ -25,7 +25,6 @@ from collections.abc import AsyncGenerator
 from collections.abc import Iterable
 from dataclasses import dataclass
 from dataclasses import field
-from datetime import UTC
 from datetime import datetime
 from datetime import timedelta
 import logging
@@ -37,15 +36,19 @@ from nat.builder.framework_enum import LLMFrameworkEnum
 from nat.builder.function_info import FunctionInfo
 from nat.cli.register_workflow import register_function
 from nat.data_models.function import FunctionBaseConfig
-from pydantic import AwareDatetime
 from pydantic import BaseModel
 from pydantic import ConfigDict
 from pydantic import Field
-from pydantic import field_validator
+
+from vss_agents.data_models.ranking import DEFAULT_CHUNK_SECONDS
+from vss_agents.data_models.ranking import ChunkKey
+from vss_agents.data_models.ranking import RankedChunk
+from vss_agents.data_models.ranking import RankedList
+from vss_agents.data_models.ranking import _validate_chunk_seconds
+from vss_agents.data_models.ranking import snap
 
 FusionMethod = Literal["rrf", "weighted_linear"]
 Aggregation = Literal["max", "mean"]
-DEFAULT_CHUNK_SECONDS = 5
 DEFAULT_RRF_K = 60
 
 FiniteFloat = Annotated[float, Field(allow_inf_nan=False)]
@@ -55,69 +58,8 @@ FiniteNonNegFloat = Annotated[float, Field(ge=0, allow_inf_nan=False)]
 logger = logging.getLogger(__name__)
 
 # ---------------------------------------------------------------------------
-# Public data contract
+# Fusion data contract
 # ---------------------------------------------------------------------------
-
-
-class ChunkKey(BaseModel):
-    """Unique identifier for a 5-second video chunk on the snapped grid.
-
-    Note: End of the chunk is intentionally not a field of the key. End is fully derived
-    as ``start + chunk_seconds`` post-bucketize.
-    """
-
-    # Makes model hashable so it can be used as a ``dict`` key during the outer-join
-    model_config = ConfigDict(frozen=True)
-
-    sensor_id: str
-    start: AwareDatetime = Field(
-        description=(
-            "Raw upstream timestamp; bucketize snaps to chunk_seconds grid. "
-            "Must be tz-aware (naive rejected; non-UTC coerced to UTC)."
-        ),
-    )
-
-    @field_validator("start")
-    @classmethod
-    def _coerce_to_utc(cls, v: datetime) -> datetime:
-        """Normalize any tz-aware datetime to UTC.
-
-        Note: ``AwareDatetime`` has already rejected naive non-tz inputs at this point.
-        """
-        return v.astimezone(UTC)
-
-
-class RankedChunk(BaseModel):
-    """One ranked entry from a single embedding space.
-
-    Pure data model. Fusion is blind to payloads.
-    Hence no VST URLs, screenshots, descriptions, object_ids, they live on the
-    original search results in ``search.py``.
-    """
-
-    model_config = ConfigDict(frozen=True)
-
-    key: ChunkKey
-
-    # Raw score in space-native units (cosine, frame_score, …)
-    score: float = Field(allow_inf_nan=False)
-
-    # 1-based position inside its source list
-    # Note: Fusion does not compute initial ranks - it receives them
-    # (the ranks come baked into the input from upstream search tools)
-    # e.g. each embedding space tool -> emits ``RankedList(rank=1..K)``
-    rank: int
-
-
-class RankedList(BaseModel):
-    """A ranked list of chunks from one embedding space."""
-
-    model_config = ConfigDict(frozen=True)
-
-    # Kept dynamic (not a literal) for ease of extensibility when consumers add new spaces
-    # "embed", "attribute", "caption", "face", ...
-    space: str
-    chunks: list[RankedChunk] = Field(default_factory=list)
 
 
 class _SharedFusionParams(BaseModel):
@@ -293,40 +235,10 @@ class FusedRow:
 # ---------------------------------------------------------------------------
 
 
-def _validate_chunk_seconds(chunk_seconds: int) -> None:
-    """Invariant for callers that bypass :class:`FusionInput`."""
-    if chunk_seconds <= 0:
-        raise ValueError(f"chunk_seconds must be > 0, got {chunk_seconds!r}")
-
-
 def _validate_rrf_k(rrf_k: int) -> None:
     """Invariant for callers that bypass :class:`FusionInput`."""
     if rrf_k <= 0:
         raise ValueError(f"rrf_k must be > 0, got {rrf_k!r}")
-
-
-# TODO consumed by ``search.py`` for adapters.
-# When other embedding tools are refactored to return ranked lists natively, move this
-# and the data models above (ChunkKey, RankedList, etc.) to ``data_models/ranking.py``.
-def snap(ts: AwareDatetime, chunk_seconds: int = DEFAULT_CHUNK_SECONDS) -> datetime:
-    """Snap an arbitrary timestamp down to the chunk-grid floor.
-
-    Different search tools may return chunks at slightly different timestamps
-    even when describing the same moment of video. Snapping lines them up onto a deterministic grid
-    so the outer-join in :func:`fuse` matches the same chunk across spaces.
-
-    Example:
-    - `embed_search` (Cosmos clip embeddings) -> 00:01:25.300 (sliding window started)
-    - `attribute_search` (CV per-frame) -> 00:01:27.100 (bounding box landed)
-    -> both snap to 00:01:25
-    """
-    _validate_chunk_seconds(chunk_seconds)
-    if ts.tzinfo is None:
-        raise ValueError(f"snap requires a tz-aware datetime, got naive {ts!r}")
-    epoch = datetime(1970, 1, 1, tzinfo=ts.tzinfo)
-    seconds_since_epoch = (ts - epoch).total_seconds()
-    snapped = (seconds_since_epoch // chunk_seconds) * chunk_seconds
-    return epoch + timedelta(seconds=snapped)
 
 
 def _score_with_tiebreak(score: float, sensor_id: str, start: datetime) -> tuple[float, str, datetime]:
